reorganisation in folders to symbolize components
