+++ UDV-Core/src/Modules/Contribute/DocumentsBrowser.js
@@ -3,17 +3,20 @@
 * Description :
 * The Document Handler is an object holding and managing Document objects
 * It handles the display of documents in the document browser window, the central window, and billboards.
+*
 * Documents are objects with properties : source image, title, date, metadata, camera position,
 * camera quaternion (both for the oriented view) and billboard position
 */
 
 import * as THREE from 'three';
 import { MAIN_LOOP_EVENTS } from 'itowns';
-import { Document } from './Document.js'
-import { readCSVFile } from '../../Tools/CSVLoader.js';
-import './DocumentsHandler.css';
+import { ExtendedDocument } from './ExtendedDocument.js'
+import './Contribute.css';
 import DefaultImage from './DefaultImage.png';
-
+import { CreateDoc } from './CreateDoc.js';
+import { DeleteDoc } from './DeleteDoc.js';
+import { UpdateDoc } from './UpdateDoc.js';
+//import {CreateDoc } from './CreateDoc.js';
 
 // TO DO : pass showBillboardButton as an option to DocumentsHandler
 // currently, BILLBOARDS WILL BE ALWAYS HIDDEN if the showBillboardButton global var is set to false !!
@@ -29,7 +32,7 @@ import DefaultImage from './DefaultImage.png';
 * @param options : optional parameters (including TemporalController)
 */
 //=============================================================================
-export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {}) {
+export function DocumentsBrowser(view, controls, jsonDataFromDB, options = {}) {
 
     ///////////// Html elements
     var docDiv = document.createElement("div");
@@ -38,6 +41,8 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
 
     document.getElementById("doc").innerHTML =
     '<div id="docBrowserWindow">\
+      <button id="closeBrowserWindow" type=button>X</button><br/>\
+      <br/>\
         <div id="docHead">Document Navigator</div>\
         <div id="docBrowserTitle">doc title</div>\
         <div id="docBrowserMetaData">metadata</div>\
@@ -46,8 +51,12 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
         <div id="docBrowserIndex"></div>\
         <button id="docBrowserNextButton" type=button>⇨</button>\
         <button id="docBrowserPreviousButton" type=button>⇦</button>\
+        <div id="operationsOnDoc">\
+        <button id="docDelete" type = button>Delete</button>\
+        <button id = "docUpdate" type = button>Update</button>\
         <button id="docBrowserOrientButton" type=button>Orient Document</button>\
-        <button id = "docUpdateBIS" type = button>Update Bis</button>\
+        </div>\
+        <button id = "docCreateFromBrowser" type = button>Create new doc</button>\
     </div>\
     <div id="docFull">\
         <img id="docFullImg"/>\
@@ -65,9 +74,28 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     style="display:none;">Billboard</button>\
     ';
 
-    /////// Class attributes
+    var updateDiv = document.createElement("div");
+    updateDiv. id = "UpdateWindow";
+    document.body.appendChild(updateDiv);
+    document.getElementById("UpdateWindow").innerHTML =
+    '<div id = "updateDocWindow">\
+    <div id = updateDocForm></div>\
+    <img id="filePreview"/>\
+    <button id = "saveUpdateButton" type = "button">Save</button>\
+    <button id = "closeUpdateForm" type="button">Cancel</button>\
+    </div>\
+    ';
 
-    // Whether the Document Handler sub window displaying controlling GUI elements
+    var schema = "http://rict.liris.cnrs.fr/schemaType.json";
+    var optionsUpdate = "http://rict.liris.cnrs.fr/optionsUpdate.json  ";
+
+    $("#updateDocForm").alpaca({
+         "schemaSource": schema,
+         "optionsSource": optionsUpdate
+});
+
+    /////// Class attributes
+    // Whandleether the Document Handler sub window displaying controlling GUI elements
     // is currently displayed or not.
     this.docBrowserWindowIsActive = options.docBrowserWindowStartActive || false;
 
@@ -95,7 +123,7 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     this.camera = view.camera.camera3D;
 
     // path to the csv file holding the guided tour data
-    const CSVdataFile = dataFile;
+    //const CSVdataFile = dataFile;
 
     // TemporalController instance (optional)
     // this is used to set the current date according to the selected document
@@ -126,56 +154,55 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
 
     /**
     * initialize the controller using data from the csv file
-    * this function is called after the completion of readCSVFile() in this.loadDataFromFile()
-    * @param docDataFromFile : contains the data loaded from the file
+    * this function is called after the completion of readCSVFile() in this.loadDataFromDatabase()
+    * @param docDataFromDB : contains the data loaded from the database
     */
     //==========================================================================
-    this.initialize = function initialize(docDataFromFile){
-//console.log("ini");
+    this.initialize = function initialize(docDataFromDB){
+      console.log(docDataFromDB);
         // fill the AllDocuments array with Documents objects
         // the Documents are placed in the order they are loaded, which is their line order in the csv file
         // the docIndex property is specified to be 0,1,2,3 etc... in the csv
         // therefore docIndex is equal to "i", but we specify it in the csv for clarity (we need docIndex for the guided tour csv)
         // the difference between docIndex and doc_ID (used by historians) should be settled asap
-        for (var i=0; i<docDataFromFile.length; i++) {
+        var url = "http://127.0.0.1/APIExtendedDocument/web/documentsDirectory/";
+//        var url = "http://rict.liris.cnrs.fr/APIVilo3D/APIExtendedDocument/web/documentsDirectory/";
+        for (var i=0; i<docDataFromDB.length; i++) {
 
-            var docData = docDataFromFile[i];
-            var docIndex = parseFloat(docData[0]);
-            var doc_ID = parseFloat(docData[1]);
-            //this is a test to use the images stored in the RICT server and not in UDV
-
-            //var docImageSourceHD = "Vilo3D/Docs/"+docData[2];
-            //var docImageSourceBD = "Vilo3D/Docs/"+docData[3];
-            var docImageSourceHD = "http://rict.liris.cnrs.fr/DataStore/Vilo3Ddocs/"+docData[2];
-            console.log("coucou");
-            var docImageSourceBD = "http://rict.liris.cnrs.fr/DataStore/Vilo3Ddocs/"+docData[3];
-            var docTitle = docData[4].toString();
-
-            var docStartDate = new moment( docData[5].toString() );
-
-            var docMetaData = docData[6].toString();
+            var docData = docDataFromDB[i];
+            var docIndex = i;
+            var doc_ID = docData.idDocument
+            var docImageSourceHD =  url + docData.metadata.link;
+            var docImageSourceBD = url + docData.metadata.link;
+            var docTitle = docData.metadata.title;
+            var docDescription = docData.metadata.description;
+            //var docStartDate = moment('2016-01-01');
+            var docRefDate = docData.metadata.refDate;
+            var docPublicationDate = docData.metadata.publicationDate;
+            var docSubject = docData.metadata.subject;
+            //var docRefDate = moment('2016-01-01');
+            var docMetaData = "Referring date: " + docRefDate +" Publication date: " + docPublicationDate;
 
             // camera position for the oriented view
             var docViewPos = new THREE.Vector3();
-            docViewPos.x = parseFloat(docData[7]);
-            docViewPos.y = parseFloat(docData[8]);
-            docViewPos.z = parseFloat(docData[9]);
+            docViewPos.x = docData.visualization.positionX;
+            docViewPos.y = docData.visualization.positionY;
+            docViewPos.z = docData.visualization.positionZ;
 
             // camera orientation for the oriented view
             var docViewQuat = new THREE.Quaternion();
-            docViewQuat.x = parseFloat(docData[10]);
-            docViewQuat.y = parseFloat(docData[11]);
-            docViewQuat.z = parseFloat(docData[12]);
-            docViewQuat.w = parseFloat(docData[13]);
+            docViewQuat.x = docData.visualization.quaternionX;
+            docViewQuat.y = docData.visualization.quaternionX;
+            docViewQuat.z = docData.visualization.quaternionX;
+            docViewQuat.w = docData.visualization.quaternionX;
 
             // billboard position
             var docBillboardPos = new THREE.Vector3();
-            docBillboardPos.x = parseFloat(docData[14]);
-            docBillboardPos.y = parseFloat(docData[15]);
-            docBillboardPos.z = parseFloat(docData[16]);
+            docBillboardPos.x = 1;
+            docBillboardPos.y = 1;
+            docBillboardPos.z = 1;
 
-            var doc = new Document(docTitle,docIndex,doc_ID,docImageSourceHD,docImageSourceBD,docBillboardPos,docViewPos,docViewQuat,docStartDate,docMetaData);
-
+            var doc = new ExtendedDocument(docTitle,docIndex,doc_ID,docImageSourceHD,docImageSourceBD,docBillboardPos,docViewPos,docViewQuat,docRefDate, docPublicationDate,docDescription, docMetaData, docSubject);
             // we fill the AllDocuments array with the new doc
             // this doc is accessed using AllDocuments[docIndex]
             this.AllDocuments.push(doc);
@@ -184,9 +211,9 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
 
         // load the first doc as current doc
         this.currentDoc = this.AllDocuments[0];
+        //console.log(this.currentDoc);
 
         this.updateBrowser();
-
         if(billboardsAreActive){
             this.showBillboards(true);
         }
@@ -201,9 +228,10 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     }
 
     //==========================================================================
-    this.loadDataFromFile = function loadDataFromFile(){
+    this.loadDataFromDatabase = function loadDataFromDatabase(){
 
-        readCSVFile(CSVdataFile, this.initialize.bind(this));
+        this.initialize(jsonDataFromDB);
+        //readCSVFile(CSVdataFile, this.initialize.bind(this));
 
     }
 
@@ -280,6 +308,7 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
         else {
 
             this.currentDoc = this.AllDocuments[index+1];
+            //console.log(this.currentDoc.getDocID());
             this.updateBrowser();
         }
     }
@@ -304,9 +333,11 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     this.updateBrowser = function updateBrowser(){
 
         document.getElementById('docBrowserPreviewImg').src = this.currentDoc.imageSourceBD;
-        document.getElementById('docBrowserMetaData').innerHTML = this.currentDoc.metaData;
+        document.getElementById('docBrowserMetaData').innerHTML = this.currentDoc.metadata;
         document.getElementById('docBrowserTitle').innerHTML = this.currentDoc.title;
-        document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.index;
+        document.getElementById('docDescription').innerHTML = this.currentDoc.description;
+        //document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.index;
+      //  document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.doc_ID;
     }
 
     // show billboards
@@ -314,6 +345,7 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     // if forceShow is false and billboardsAreActive is false, do not show
     //==========================================================================
     this.showBillboards = function showBillboards(forceShow){
+      console.log('showBillboards');
 
         if(!forceShow && !billboardsAreActive){
             return;
@@ -375,6 +407,25 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
         }
     }
 
+    this.handleDocUpdate = function handleDocUpdate(){
+      //show window
+      console.log("test update");
+      //console.log(this.currentDoc.doc_ID);
+      var Update = new UpdateDoc(this.currentDoc);
+
+    }
+
+    this.handleDocDelete = function handleDocDelete(){
+      console.log('doc deletion');
+      var Delete = new DeleteDoc(this.currentDoc);
+    }
+
+    this.handleDocCreation = function handleDocCreation(){
+      console.log("doc creation");
+      var Create = new CreateDoc(this.controls, this.view);
+    }
+
+
     // triggers the "oriented view" of the current docIndex
     // this will display the doc image in the middle of the screen
     // and initiate the animated travel to orient the camera
@@ -392,13 +443,14 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
 
         // if we have valid data, initiate the animated travel to orient the camera
         if(!isNaN(this.currentDoc.viewPosition.x) && !isNaN(this.currentDoc.viewQuaternion.x)){
+          //console.log(this.currentDoc.viewPosition );
+          //console.log(this.currentDoc.viewQuaternion);
           this.controls.initiateTravel(this.currentDoc.viewPosition,"auto",this.currentDoc.viewQuaternion,true);
         }
 
         // adjust the current date if we use temporal
         if(this.temporal){
-
-            temporal.changeTime(this.currentDoc.startDate);
+        //  temporal.changeTime(this.currentDoc.refDate1);
         }
 
         this.hideBillboards(true);
@@ -411,10 +463,6 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
         this.view.camera
 
     };
-
-    this.handleDocUpdate = function handleDocUpdate(){
-      console.log(this.currentDoc.doc_ID);
-    }
 
     // close the center window (oriented view / doc focus)
     //=========================================================================
@@ -487,15 +535,17 @@ export function DocumentsHandlerEmmanuel(view, controls, dataFile, options = {})
     document.getElementById("docBrowserNextButton").addEventListener('mousedown',this.nextDoc.bind(this),false);
     document.getElementById("docBrowserPreviousButton").addEventListener('mousedown',this.previousDoc.bind(this),false);
     document.getElementById("docBrowserOrientButton").addEventListener('mousedown', this.focusOnDoc.bind(this),false);
-    document.getElementById("docUpdateBIS").addEventListener('mousedown', this.handleDocUpdate.bind(this),false);
+    document.getElementById("docUpdate").addEventListener('mousedown', this.handleDocUpdate.bind(this),false);
+    document.getElementById("docDelete").addEventListener('mousedown', this.handleDocDelete.bind(this),false);
+    document.getElementById("docCreateFromBrowser").addEventListener('mousedown', this.handleDocCreation.bind(this),false);
+
 
     // setup display
     document.getElementById("docBrowserToggleBillboard").style.display = (showBillboardButton)? "block" : "none";
     document.getElementById("docBrowserWindow").style.display = (!this.docBrowserWindowIsActive)? "none" : "block";
 
-
     // this will trigger the initialization, after file loading is complete
-    this.loadDataFromFile();
+    this.loadDataFromDatabase();
 
 }
