+++ UDV-Core/src/Modules/Documents/DocumentsHandler.css
@@ -23,7 +23,7 @@
     background-color: rgba(30,30,30,0.6);
     font-size: 14;
 
-    /*pointer-events: none;*/
+    pointer-events: none;
 }
 
 #docHead{
@@ -45,28 +45,23 @@
     font-size: 16px;
     font-weight: bold;
     overflow: hidden;
-    width:95%
 }
 
 #docBrowserMetaData{
-  display:inline-block;
-  margin-top: 20px;
-  color: black;
-  background-color: ivory;
-  text-align: center;
-  font-size: 12px;
-  overflow: hidden;
-  width:95%
-}
-#docDescription{
-  display:inline-block;
-  margin-top: 20px;
-  color: black;
-  background-color: ivory;
-  text-align: center;
-  font-size: 12px;
-  overflow: hidden;
-  width:95%
+    display:inline-block;
+    margin-top: 20px;
+    margin: 10px;
+    padding: 5px;
+    padding-left: 10px;
+    padding-right: 10px;
+    line-height: 130%;
+    height: 10%;
+    color : black;
+    background-color: ivory ;
+    border : 1px solid rgb(90,90,90);
+    text-align: justify;
+    overflow: hidden;
+    pointer-events: auto;
 }
 
 #docBrowserPreview{
@@ -122,36 +117,19 @@
     pointer-events: auto;
 }
 
-#saveUpdateButton{
+#docBrowserOrientButton{
     position: absolute;
     z-index: 20;
     bottom : 0;
-    left : 0;
+    right : 0;
     margin : 5px;
-    font-size: 20px;
+    font-size: 14px;
     border: 1px solid black;
     background-color: white;
     color: black;
-    height: 40px;
+    height: 30px;
     pointer-events: auto;
 }
-
-#closeUpdateForm{
-    position: absolute;
-    z-index: 20;
-    bottom : 0;
-    right: 0;
-    margin : 5px;
-    font-size: 20px;
-    border: 1px solid black;
-    background-color: white;
-    color: black;
-    height: 40px;
-    pointer-events: auto;
-}
-
-
-
 
 #docFull{
     position: absolute;
@@ -182,31 +160,6 @@
     opacity: 0.5;
 }
 
-#filePreview{
-    width: 80%;
-    height: 40%;
-    object-fit: contain;
-}
-
-#updateDocWindow{
-    display: none;
-    position: absolute;
-    z-index: 10;
-    margin : 1px;
-    margin-right: 1px;
-    top: 1px;
-    right: 0;
-    border : 1px solid rgb(90,90,90);
-    text-align: center;
-    color: white;
-    width: 22%;
-    height: 75%;
-    color : black;
-    background-color: rgba(30,30,30,0.6);
-    font-size: 14;
-}
-
-
 #docFullPanel{
     width: 60%;
     height: 50px;
@@ -225,13 +178,19 @@
 #docFullClose{
     position: absolute;
     z-index: 10;
+
     margin : auto;
+
     bottom : 20%;
+
     left: 2%;
+
     border: 1px solid black;
     background-color: white;
     color: black;
+
     height: 30px;
+
     pointer-events: auto;
 }
 
@@ -311,74 +270,4 @@
     right: 15%;
 
     pointer-events: auto;
-}
-
-#operationsOnDoc{
-  margin: 10px;
-  /*  border: 3px solid green;*/
-
-}
-
- #docDelete{
-
-  font-size: 14px;
-  border: 1px solid black;
-  background-color: white;
-  color: black;
-  height: 30px;
-  pointer-events: auto;
-}
-
-#docUpdate{
-font-size: 14px;
-  border: 1px solid black;
-  background-color: white;
-  color: black;
-  height: 30px;
-  pointer-events: auto;
-}
-#closeBrowserWindow{
-  position: absolute;
-  z-index: 20;
-  top : 0;
-  right : 0;
-  margin : 5px;
-  font-size: 14px;
-  border: 1px solid black;
-  background-color: white;
-  color: black;
-  height: 30px;
-  pointer-events: auto;
-
-}
-
-#docBrowserOrientButton{
-    font-size: 14px;
-    border: 1px solid black;
-    background-color: white;
-    color: black;
-    height: 30px;
-    pointer-events: auto;
-}
-
-#docCreateFromBrowser{
-  position: absolute;
-  z-index: 20;
-  right : 0;
-      bottom : 0;
-  margin : 5px;
-  font-size: 14px;
-  border: 2px solid #008CBA;
-  background-color: white;
-  color: black;
-  height: 30px;
-  pointer-events: auto;
-}
-
-#debugCameraPostion{
-  display: none;
-}
-
-#setPositionFields{
-  display: none;
 }
