+++ UDV-Core/src/Modules/Documents/DocumentsHandler.js
@@ -3,7 +3,6 @@
 * Description :
 * The Document Handler is an object holding and managing Document objects
 * It handles the display of documents in the document browser window, the central window, and billboards.
-*
 * Documents are objects with properties : source image, title, date, metadata, camera position,
 * camera quaternion (both for the oriented view) and billboard position
 */
@@ -14,8 +13,6 @@ import { Document } from './Document.js'
 import { readCSVFile } from '../../Tools/CSVLoader.js';
 import './DocumentsHandler.css';
 import DefaultImage from './DefaultImage.png';
-import '../Contribute/DocumentPositioner.js';
-
 
 // TO DO : pass showBillboardButton as an option to DocumentsHandler
 // currently, BILLBOARDS WILL BE ALWAYS HIDDEN if the showBillboardButton global var is set to false !!
@@ -31,7 +28,7 @@ import '../Contribute/DocumentPositioner.js';
 * @param options : optional parameters (including TemporalController)
 */
 //=============================================================================
-export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
+export function DocumentsHandler(view, controls, dataFile, options = {}) {
 
     ///////////// Html elements
     var docDiv = document.createElement("div");
@@ -40,8 +37,6 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
 
     document.getElementById("doc").innerHTML =
     '<div id="docBrowserWindow">\
-      <button id="closeBrowserWindow" type=button>X</button><br/>\
-      <br/>\
         <div id="docHead">Document Navigator</div>\
         <div id="docBrowserTitle">doc title</div>\
         <div id="docBrowserMetaData">metadata</div>\
@@ -50,12 +45,7 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
         <div id="docBrowserIndex"></div>\
         <button id="docBrowserNextButton" type=button>⇨</button>\
         <button id="docBrowserPreviousButton" type=button>⇦</button>\
-        <div id="operationsOnDoc">\
-        <button id="docDelete" type = button>Delete</button>\
-        <button id = "docUpdate" type = button>Update</button>\
         <button id="docBrowserOrientButton" type=button>Orient Document</button>\
-        </div>\
-        <button id = "docCreateFromBrowser" type = button>Create new doc</button>\
     </div>\
     <div id="docFull">\
         <img id="docFullImg"/>\
@@ -73,31 +63,9 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     style="display:none;">Billboard</button>\
     ';
 
-    var updateDiv = document.createElement("div");
-    updateDiv. id = "UpdateWindow";
-    document.body.appendChild(updateDiv);
-    document.getElementById("UpdateWindow").innerHTML =
-    '<div id = "updateDocWindow">\
-    <div id = updateDocForm></div>\
-    <img id="filePreview"/>\
-    <button id = "saveUpdateButton" type = "button">Save</button>\
-    <button id = "closeUpdateForm" type="button">Cancel</button>\
-    </div>\
-    ';
-
-    var schema = "http://rict.liris.cnrs.fr/schemaType.json";
-    var optionsUpdate = "http://rict.liris.cnrs.fr/optionsUpdate.json  ";
-
-    $("#updateDocForm").alpaca({
-         "schemaSource": schema,
-         "optionsSource": optionsUpdate
-    });
-
-
-
-
     /////// Class attributes
-    // Whandleether the Document Handler sub window displaying controlling GUI elements
+
+    // Whether the Document Handler sub window displaying controlling GUI elements
     // is currently displayed or not.
     this.docBrowserWindowIsActive = options.docBrowserWindowStartActive || false;
 
@@ -125,7 +93,7 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     this.camera = view.camera.camera3D;
 
     // path to the csv file holding the guided tour data
-    //const CSVdataFile = dataFile;
+    const CSVdataFile = dataFile;
 
     // TemporalController instance (optional)
     // this is used to set the current date according to the selected document
@@ -156,54 +124,51 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
 
     /**
     * initialize the controller using data from the csv file
-    * this function is called after the completion of readCSVFile() in this.loadDataFromDatabase()
-    * @param docDataFromDB : contains the data loaded from the database
+    * this function is called after the completion of readCSVFile() in this.loadDataFromFile()
+    * @param docDataFromFile : contains the data loaded from the file
     */
     //==========================================================================
-    this.initialize = function initialize(docDataFromDB){
+    this.initialize = function initialize(docDataFromFile){
+
         // fill the AllDocuments array with Documents objects
         // the Documents are placed in the order they are loaded, which is their line order in the csv file
         // the docIndex property is specified to be 0,1,2,3 etc... in the csv
         // therefore docIndex is equal to "i", but we specify it in the csv for clarity (we need docIndex for the guided tour csv)
         // the difference between docIndex and doc_ID (used by historians) should be settled asap
+        for (var i=0; i<docDataFromFile.length; i++) {
 
-        var url = "http://rict.liris.cnrs.fr/APIVilo3D/APIExtendedDocument/web/documentsDirectory/";
-        for (var i=0; i<docDataFromDB.length; i++) {
+            var docData = docDataFromFile[i];
+            var docIndex = parseFloat(docData[0]);
+            var doc_ID = parseFloat(docData[1]);
+            var docImageSourceHD = "Vilo3D/Docs/"+docData[2];
+            var docImageSourceBD = "Vilo3D/Docs/"+docData[3];
+            var docTitle = docData[4].toString();
 
-            var docData = docDataFromDB[i];
-            var docIndex = i;
-            var doc_ID = docData.idDocument
-            var docImageSourceHD =  url + docData.metadata.link;
-            var docImageSourceBD = url + docData.metadata.link;
-            var docTitle = docData.metadata.title;
-            var docDescription = docData.metadata.description;
-            //var docStartDate = moment('2016-01-01');
-            var docRefDate = docData.metadata.refDate;
-            var docPublicationDate = docData.metadata.publicationDate;
-            var docSubject = docData.metadata.subject;
-            //var docRefDate = moment('2016-01-01');
-            var docMetaData = "Referring date: " + docRefDate +" Publication date: " + docPublicationDate;
+            var docStartDate = new moment( docData[5].toString() );
+
+            var docMetaData = docData[6].toString();
 
             // camera position for the oriented view
             var docViewPos = new THREE.Vector3();
-            docViewPos.x = docData.visualization.positionX;
-            docViewPos.y = docData.visualization.positionY;
-            docViewPos.z = docData.visualization.positionZ;
+            docViewPos.x = parseFloat(docData[7]);
+            docViewPos.y = parseFloat(docData[8]);
+            docViewPos.z = parseFloat(docData[9]);
 
             // camera orientation for the oriented view
             var docViewQuat = new THREE.Quaternion();
-            docViewQuat.x = docData.visualization.quaternionX;
-            docViewQuat.y = docData.visualization.quaternionX;
-            docViewQuat.z = docData.visualization.quaternionX;
-            docViewQuat.w = docData.visualization.quaternionX;
+            docViewQuat.x = parseFloat(docData[10]);
+            docViewQuat.y = parseFloat(docData[11]);
+            docViewQuat.z = parseFloat(docData[12]);
+            docViewQuat.w = parseFloat(docData[13]);
 
             // billboard position
             var docBillboardPos = new THREE.Vector3();
-            docBillboardPos.x = 1;
-            docBillboardPos.y = 1;
-            docBillboardPos.z = 1;
+            docBillboardPos.x = parseFloat(docData[14]);
+            docBillboardPos.y = parseFloat(docData[15]);
+            docBillboardPos.z = parseFloat(docData[16]);
 
-            var doc = new Document(docTitle,docIndex,doc_ID,docImageSourceHD,docImageSourceBD,docBillboardPos,docViewPos,docViewQuat,docRefDate, docPublicationDate,docDescription, docMetaData, docSubject);
+            var doc = new Document(docTitle,docIndex,doc_ID,docImageSourceHD,docImageSourceBD,docBillboardPos,docViewPos,docViewQuat,docStartDate,docMetaData);
+
             // we fill the AllDocuments array with the new doc
             // this doc is accessed using AllDocuments[docIndex]
             this.AllDocuments.push(doc);
@@ -212,9 +177,9 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
 
         // load the first doc as current doc
         this.currentDoc = this.AllDocuments[0];
-        //console.log(this.currentDoc);
 
         this.updateBrowser();
+
         if(billboardsAreActive){
             this.showBillboards(true);
         }
@@ -229,10 +194,9 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     }
 
     //==========================================================================
-    this.loadDataFromDatabase = function loadDataFromDatabase(){
+    this.loadDataFromFile = function loadDataFromFile(){
 
-        this.initialize(jsonDataFromDB);
-        //readCSVFile(CSVdataFile, this.initialize.bind(this));
+        readCSVFile(CSVdataFile, this.initialize.bind(this));
 
     }
 
@@ -309,7 +273,6 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
         else {
 
             this.currentDoc = this.AllDocuments[index+1];
-            //console.log(this.currentDoc.getDocID());
             this.updateBrowser();
         }
     }
@@ -334,11 +297,9 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     this.updateBrowser = function updateBrowser(){
 
         document.getElementById('docBrowserPreviewImg').src = this.currentDoc.imageSourceBD;
-        document.getElementById('docBrowserMetaData').innerHTML = this.currentDoc.metadata;
+        document.getElementById('docBrowserMetaData').innerHTML = this.currentDoc.metaData;
         document.getElementById('docBrowserTitle').innerHTML = this.currentDoc.title;
-        document.getElementById('docDescription').innerHTML = this.currentDoc.description;
-        //document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.index;
-      //  document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.doc_ID;
+        document.getElementById('docBrowserIndex').innerHTML = "index : " + this.currentDoc.index;
     }
 
     // show billboards
@@ -346,7 +307,6 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     // if forceShow is false and billboardsAreActive is false, do not show
     //==========================================================================
     this.showBillboards = function showBillboards(forceShow){
-      console.log('showBillboards');
 
         if(!forceShow && !billboardsAreActive){
             return;
@@ -408,25 +368,6 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
         }
     }
 
-    this.handleDocUpdate = function handleDocUpdate(){
-      //show window
-      console.log("test update");
-      //console.log(this.currentDoc.doc_ID);
-      var Update = new udvcore.UpdateDoc(this.currentDoc);
-
-    }
-
-    this.handleDocDelete = function handleDocDelete(){
-      console.log('doc deletion');
-      var Delete = new udvcore.DeleteDoc(this.currentDoc);
-    }
-
-    this.handleDocCreation = function handleDocCreation(){
-      console.log("doc creation");
-      var Create = new udvcore.CreateDoc(this.controls, this.view);
-    }
-
-
     // triggers the "oriented view" of the current docIndex
     // this will display the doc image in the middle of the screen
     // and initiate the animated travel to orient the camera
@@ -444,14 +385,14 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
 
         // if we have valid data, initiate the animated travel to orient the camera
         if(!isNaN(this.currentDoc.viewPosition.x) && !isNaN(this.currentDoc.viewQuaternion.x)){
-          //console.log(this.currentDoc.viewPosition );
-          //console.log(this.currentDoc.viewQuaternion);
-          this.controls.initiateTravel(this.currentDoc.viewPosition,"auto",this.currentDoc.viewQuaternion,true);
+
+            this.controls.initiateTravel(this.currentDoc.viewPosition,"auto",this.currentDoc.viewQuaternion,true);
         }
 
         // adjust the current date if we use temporal
         if(this.temporal){
-        //  temporal.changeTime(this.currentDoc.refDate1);
+
+            temporal.changeTime(this.currentDoc.startDate);
         }
 
         this.hideBillboards(true);
@@ -461,7 +402,6 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
 
         //to request an update
         this.view.notifyChange(false);
-        this.view.camera
 
     };
 
@@ -536,17 +476,13 @@ export function DocumentsHandler(view, controls, jsonDataFromDB, options = {}) {
     document.getElementById("docBrowserNextButton").addEventListener('mousedown',this.nextDoc.bind(this),false);
     document.getElementById("docBrowserPreviousButton").addEventListener('mousedown',this.previousDoc.bind(this),false);
     document.getElementById("docBrowserOrientButton").addEventListener('mousedown', this.focusOnDoc.bind(this),false);
-    document.getElementById("docUpdate").addEventListener('mousedown', this.handleDocUpdate.bind(this),false);
-    document.getElementById("docDelete").addEventListener('mousedown', this.handleDocDelete.bind(this),false);
-    document.getElementById("docCreateFromBrowser").addEventListener('mousedown', this.handleDocCreation.bind(this),false);
-
 
     // setup display
     document.getElementById("docBrowserToggleBillboard").style.display = (showBillboardButton)? "block" : "none";
     document.getElementById("docBrowserWindow").style.display = (!this.docBrowserWindowIsActive)? "none" : "block";
 
     // this will trigger the initialization, after file loading is complete
-    this.loadDataFromDatabase();
+    this.loadDataFromFile();
 
 }
